+++ jwql/website/apps/jwql/monitor_models/bad_pixel.py
@@ -29,7 +29,7 @@ References
 from django.contrib.postgres.fields import ArrayField
 from django.db import models
 
-from jwql.utils.constants import MAX_LEN_APERTURE, MAX_LEN_DETECTOR, MAX_LEN_FILENAME, MAX_LEN_INSTRUMENT, MAX_LEN_TYPE
+from jwql.utils.constants import MAX_LEN_APERTURE, MAX_LEN_DETECTOR, MAX_LEN_FILENAME, MAX_LEN_INSTRUMENT, MAX_LEN_TYPE, DEFAULT_MODEL_CHARFIELD
 
 
 class FGSBadPixelQueryHistory(models.Model):
@@ -58,7 +58,7 @@ class FGSBadPixelStats(models.Model):
     x_coord = ArrayField(models.IntegerField())
     y_coord = ArrayField(models.IntegerField())
     type = models.CharField(max_length=MAX_LEN_TYPE, blank=True, null=True)
-    source_files = models.TextField(blank=True, null=True)  # This field type is a guess.
+    source_files = ArrayField(models.CharField(max_length=MAX_LEN_FILENAME, help_text="source file names", default=DEFAULT_MODEL_CHARFIELD), blank=True, null=True)
     obs_start_time = models.DateTimeField(blank=True, null=True)
     obs_mid_time = models.DateTimeField(blank=True, null=True)
     obs_end_time = models.DateTimeField(blank=True, null=True)
@@ -96,7 +96,7 @@ class MIRIBadPixelStats(models.Model):
     x_coord = ArrayField(models.IntegerField())
     y_coord = ArrayField(models.IntegerField())
     type = models.CharField(max_length=MAX_LEN_TYPE, blank=True, null=True)
-    source_files = models.TextField(blank=True, null=True)  # This field type is a guess.
+    source_files = ArrayField(models.CharField(max_length=MAX_LEN_FILENAME, help_text="source file names", default=DEFAULT_MODEL_CHARFIELD), blank=True, null=True)
     obs_start_time = models.DateTimeField(blank=True, null=True)
     obs_mid_time = models.DateTimeField(blank=True, null=True)
     obs_end_time = models.DateTimeField(blank=True, null=True)
@@ -134,7 +134,7 @@ class NIRCamBadPixelStats(models.Model):
     x_coord = ArrayField(models.IntegerField())
     y_coord = ArrayField(models.IntegerField())
     type = models.CharField(max_length=MAX_LEN_TYPE, blank=True, null=True)
-    source_files = models.TextField(blank=True, null=True)  # This field type is a guess.
+    source_files = ArrayField(models.CharField(max_length=MAX_LEN_FILENAME, help_text="source file names", default=DEFAULT_MODEL_CHARFIELD), blank=True, null=True)
     obs_start_time = models.DateTimeField(blank=True, null=True)
     obs_mid_time = models.DateTimeField(blank=True, null=True)
     obs_end_time = models.DateTimeField(blank=True, null=True)
@@ -172,7 +172,7 @@ class NIRISSBadPixelStats(models.Model):
     x_coord = ArrayField(models.IntegerField())
     y_coord = ArrayField(models.IntegerField())
     type = models.CharField(max_length=MAX_LEN_TYPE, blank=True, null=True)
-    source_files = models.TextField(blank=True, null=True)  # This field type is a guess.
+    source_files = ArrayField(models.CharField(max_length=MAX_LEN_FILENAME, help_text="source file names", default=DEFAULT_MODEL_CHARFIELD), blank=True, null=True)
     obs_start_time = models.DateTimeField(blank=True, null=True)
     obs_mid_time = models.DateTimeField(blank=True, null=True)
     obs_end_time = models.DateTimeField(blank=True, null=True)
@@ -210,7 +210,7 @@ class NIRSpecBadPixelStats(models.Model):
     x_coord = ArrayField(models.IntegerField())
     y_coord = ArrayField(models.IntegerField())
     type = models.CharField(max_length=MAX_LEN_TYPE, blank=True, null=True)
-    source_files = models.TextField(blank=True, null=True)  # This field type is a guess.
+    source_files = ArrayField(models.CharField(max_length=MAX_LEN_FILENAME, help_text="source file names", default=DEFAULT_MODEL_CHARFIELD), blank=True, null=True)
     obs_start_time = models.DateTimeField(blank=True, null=True)
     obs_mid_time = models.DateTimeField(blank=True, null=True)
     obs_end_time = models.DateTimeField(blank=True, null=True)